+++ src/commands.rs
@@ -1,5 +1,7 @@
+pub use self::filter::filter;
 pub use self::generate::generate;
 pub use self::lint::lint;
 
+pub mod filter;
 pub mod generate;
 pub mod lint;